clear interrupt flag can use sbi instruction remove unwanted type change
git-svn-id: https://openhr20.svn.sourceforge.net/svnroot/openhr20@198 e6282f77-c345-0410-9e5f-9d7f1dee61d8

+++ trunk/source/controller.c
@@ -134,7 +134,7 @@ uint8_t CTL_update(bool minute_ch, uint8_t valve) {
             if (temp>TEMP_MAX) {
                 valve = 100;
             } else {
-                valve = (int16_t)pid_Controller(calc_temp(temp),temp_average,valve);
+                valve = pid_Controller(calc_temp(temp),temp_average,valve);
             }
         } 
         PID_force_update = -1;
